Add support in the language for dependent records.
A record is basically Sigma*.
A record is paramterized by a telescope (dependent list) of types.
A tuple is a record value. Unlike the pair constructor for Sigma,
each value of a tuple also binds to a variable.
This allows for a program to be efficiently represented as a Record.

The previous elaboration of nested Sigmas and Pis simulates records,
but it results in extremely big program terms.

Also, currently the syntax for program definitions is implicitly parsed into
one big record, but the parser does not yet support an explicit
record syntax.