Correctly evaluate dependent tuples (values of dependent records).
Unfortunately, dependent records to not model programs exactly.

Consider the following program:

  myTrue : Bool
  myTrue = true

  bar : if myTrue then Bool else Unit
  bar = myTrue

  foo : if bar then Bool else Unit
  foo = bar

The type of foo is ill-typed. The if expects a Bool, but has been given
bar, which has type "if myTrue then Bool else Unit".

The same problem is evident in Agda when defining a dependent record:

  record MyProg : Set where
    field
      myTrue : Bool
      bar : if myTrue then Bool else Unit
      foo : if bar then Bool else Unit

Similarly, the following Spire nested dependent pair illustrates the problem:

  myType : Type
  myType = (myTrue : Bool) * ((bar : if myTrue then Bool else Unit) * (if bar then Bool else Unit))