Pretty print type checker output.
Not sure what's best to print here.  In the long run, we probably want
to print different ways based on command line options, but for now I
opted to interleave the pretty and "ugly" printings, so it's easy to
compare them side by side, but then to pretty print alone at the end,
in case we don't want to see all the noise of the ugly printing.