Fix minor bug in parser: 'Desc' was not a keyword, meaning it parsed as var.
I believe keywords are not supposed to be rebindable right now; it
certainly would cause problems to rebind 'caseBool'.  This was mostly
enforced in the parse, except that 'Desc' had not been added to the
keyword list.

Later, if we want keywords to be rebindable, then I think we'll need
to bind them all in an initial environment, so that they can be
treated uniformly whether they are rebound or not.