More examples.
Sum.spire fails type checking, but it shouldn't.
I think the issue is that the munger is not going inside of the pair,
and rather only updates if you subsitute a variable directly for a variable.
This leads to variable capture when substituting in a canonical value that contains
variable.