[minor] Thread identifier list through canonical embedders.
This commit does not introduce any changes in functionality, but does
change almost every line in the file. So, I factored it out from the
actual bug fix (capture avoidance).  Another option here was a reader
monad, but this version, with the explicit threading, is much easier
to effect using `M-x replace-regexp` :P