Simplify the definition of 'mkMM'.
Well, make it shorter anyway ... it's probably harder to understand
now, but the stupid second can't-possibly-fail 'gcast' is no longer
necessary now: instead of casting 'm b' to 'm a', applying the
'm a -> m a' function, and then casting 'm a' back to 'm b', we just
cast 'm a -> m a' to 'm b -> m b' once.

To make GHC understand that 'm b -> m b' has the form 'c b' for some
'c', we must define 'c = MM m'.

Note that, as shown in a comment, we could use 'ext0' define 'mkMM'
(or 'extMM', and then 'mkMM' in terms of 'extMM'), and so the MM
functions fit into the same generic pattern as T, R, Q, M, etc.