Fix bug: type checker did not evaluate early enough.
This fixes the bug exhibited by examples/Sum.spire.

@larrytheliquid: I realized a much simpler way to fix the problem;
'foldSub' is no longer used anywhere! The idea is to just substitute
the values from the environment when we come across there names,
instead of creating a DeBruijn variable and then subbing it away later
with 'foldSub'.

A potential downside: possible that this solution could lead to worse
error messages in some cases, where a term is type incorrect
*independent* of the *values* of it's parts.

For example:

  foo : Bool
  foo = <<huge term of type Bool>>

  -- Type incorrect, independent of value of foo.
  bar : String
  bar = foo

With the old buggy code, and possibly with a solution that substitutes
in environment values as late as possible, we would get an error
message like:

  foo has type Bool; expected String

whereas with the code in this commit, we would get an error like:

  <<huge term of type Bool>> has type Bool; expected String

if we display the evaluated term.  Of course, we could just as well
display the source term, instead of the evaluted term, which should
avoid the problem in this example.