Fix (inconsequential) bug: adjust free var indices.
The free variables collected by 'freeVarsDB' should
be decremented by the number of enclosing binders.
In other words, they need to be adjusted for the
level at which 'freeVarsDB' was called.  Of course,
we only use the *names* of free vars, not their
indices, so this bug wasn't causing any problems.