Elaborate wild cards.
Wild card elaboration is non trivial, because we can't simply insert
hole bindings where ever we encounter a wild card in the input
program: consider

  id : (T:Type) -> T -> T

  id' : (T:Type) -> T -> T
  id' = \ _ -> x -> id _ x

If we elaborate id' to

  id' = \ _ -> x -> id (?W:Type. ?w:W. w) x

then

  id (?W:Type. ?w:W. w)

will refine (via unification) to

  id (?w:Type. w) : (?w:Type. w) -> (?w:Type. w)

which is wrong, since the domain and range must be the same, but the
connection between them has been lost by creating two fresh, unrelated
mvars (both called "w").  Rather, we want

  ?w:Type. id w : w -> w

and for refinement to use the type 'T' of 'x' to solve 'w' by 'T'.

Note that this also means we don't want to push the wild card's mvar
binding outside the lambda, because then 'T' would not be in scope.
So, we elaborate to

  id' = \ _ -> x -> (?W:Type. ?w:W. id w x)

and not

  id' = ?W:Type. ?w:W. \ _ -> x -> id w x .

Another possibility is to lift the mvars generated for wild cards all
the way up.  This seems to require domain annotations on our lambdas,
and so we imagine elaborating from

  id' : (T:Type) -> T -> T
  id' = \ T:Type -> x:T -> id _ x

to

  id' = ?W:(T:Type) -> T -> Type. ?w:(T:Type) -> (x:T) -> W T x.
        \ T:Type -> x:T -> id (w T x) x .

This seems closer in spirit to delayed mvar substitutions in the
Matita elaboration ...