Make def of 'isWildcard' more standard.
I.e., only recognize the string "_" to be a wildcard.  Note that
'name2String' projects out the string part of a 'Name', so that, e.g.,
a wildcard that *prints* as "_3" because of freshening will still be
recognized as a wildcard here.

The 'isWildcard' function is only used in the pretty printer, so this
change shouldn't affect anything.