Make pretty printing of names slightly more legible.
The "fix" I introduced earlier, i.e. to print the
freshenings, is really annoying in practice.  So, I've
inserted a "$" between the name and the freshening
index to make the index a little easier to ignore.  But,
it would be better to only print the index when necessary ...
I'll try that next.