Pretty print fewer unnecessary freshenings.
A while back I "corrected" the pretty printer,
but the result was too ugly.  This is not perfect,
but it's better: we only print a freshening if there
is another variable in scope, with the same name, at the
binding site.

None of this really matters unless we are printing
evaluated terms ...