Define 'forcePi', up to unification related 'undefined's.
The 'forcePi' is quite complicated and so I'm not super confident it's
correct ... would be nice if I could easily test it.  Right now, I
only expect to introduce trivial unification problems
(rigid-flexible), and so one option would be to write a baby unifier
before porting the full Gundry unifier.