Add 'Spire.Debug' and use it to optionally print debug messages.
I've had uncommitted debug code littered throughout the checker for a
while, which makes checking what's changed annoying.  So, I've changed
the code to use the --debug flag to optionally enable the printing of
these messages.  I summarize some ideas for a more sophisticated
approach in the notes.org.

I made 'Spire.Debug.debug' flip the arg order relative to
'Debug.Trace.trace', which supports writing

  <computation> `debug` <msg for <computation>>

If I later make a safer (monadic) debug logger, as described in
notes.org, e.g.

  log :: a -> LogData -> SpireM a

then the usage might be

  mx >>= flip log <log data>

for 'mx :: SpireM a' and

  x `log` <log data>

for 'x :: a'.