Add Nat type.
A lot of the next few added types will be superflous once Desc'riptions are there.
However, while we're in the middle of getting that stuff working its easier to reason
about and write examples for non-encoded datatypes. This will be especially true when
implementing dependent pattern matching elaboration, as the existing work describes how
to do this in terms of normal eliminators (induction principles)
rather than the special "ind" eliminator for Mu/Desc.