CP: failed simplification of 'forcePi'.
In anticipation of extending 'forcePi' to support mvars applied to
mvars, I'm trying to stop specifying types of generated mvars.  The
mvars applied to mvars comes up in this example:

  id _ id _ x

The hope was that I didn't have to build the types of the mvars
introduced by 'forcePi', and that instead I could just have
unification solve for them.  This change implements that hope, but it
seems it does not work: I get errors from the Gundry code about

  Type does not make Type equal to ?mvar

during validation.

Now, I might be able to fix this on the Gundry end by adding
unification to the validator -- e.g. the validator could call the
unifier recursively when it encounters a unification problem when
checking judgmental equality (what I think is happening here) -- but
it's not completely obvious that would be sound.

For now, as a compromise, I will next try to construct a skeletal type
in 'forcePi' for introduced mvars, and see if that's sufficient.