CP: another failing version, but this time I think an old bug is to blame!
Here I use the type of the 'forcePi'd mvar '?T' to infer a type for
the introduced mvars '?A' and '?B' in

  ?T e1 ... en = Pi x:?A e1 ... en. ?B e1 ... en x

Namely, if if '?T : Tt' then '?A : Tt' and if

  Tt = Pi x1:T1. .... xn:Tn . Type

then 'B' has type

       Pi x1:T1. .... xn:Tn . ?A x1 ... xn -> Type

Of course, this all depends on 'Tt' being sufficiently determined at
this point in the type checking ... I haven't convinced myself it
always should be, but this strategy works for my simple examples.

The current commit still doesn't work on all examples, but it seems
that is due to an old bug: the 'unSpine' function returns the list in
reverse of how I use it!  I'll fix that next.