Fix bug in 'forceMVApp': arguments returned in reverse order.
This is a little scary because this bug is from the first
implementation of 'forceMVApp' three months ago:
e73d8e60c0804e796a662caf4e57f3051395fe5a . However, the bug was not an
issue until now, since originally I was inferring the types of the
introduced mvars via their args, and so reversing the arg order was
not a problem (and not problem in equations).

For example, I was turning

  ?M x y

into

  Pi z : ?A y x . ?B y x z

and inferring the types of '?A' and '?B' from the types of the
variables 'x' and 'y'.  The problem comes up now because I'm trying to
infer the types of '?A' and '?B' from the type of '?M', where I assume
the argument order is the same.

I should still work to infer the types of the args directly -- even if
they aren't vars anymore -- via type inference, but I haven't tried
that yet.