Back to conversative model.
Reverting model to use obvious positivity and termination arguments.
This has the advantage of being able to have alternative canonical type theory
backends, such as Agda and Idris. The tradeoff is that our own implementation
of the canonical type theory needs to do more work, because we do not
use levitation.