Darkwing Duck!
Starting to model the One-True-Data-Declaration(TM).

Here datatype parameters are encoded as first-class telescopes,
similar to descriptions. Indices are a function from the
interpretation of a parameters telescope to an indices telescope.

First-class parameters allow for them to be treaty specially
when defining generic constructors, interpreting the parametric
constructor arguments as implicit arguments.

Subsequent commits should extend this development with my generic
eliminator construction, where parameters also become implicit arguments
for the generic eliminator.

Additionally, this development makes it obvious that we also need
a generic type former, to complement our generic constructors and eliminators.